requirements.txt
# Minimum Django and REST framework version
Django>=1.11
djangorestframework>=2.4.3

# Test requirements
pytest-django>=3.4.8
pytest>=4.4.1
pytest-cov>=2.6.1
flake8==2.2.2

# wheel for PyPI installs
wheel==0.24.0

# MkDocs for documentation previews/deploys
mkdocs==0.11.1
